+++ adapters/Chartboost/CHANGELOG.md
@@ -1,8 +1,13 @@
 ## Chartboost iOS Mediation Adapter Changelog
 
-#### Next version
-- Replaced the deprecated method GADMobileAds.sharedInstance.sdkVersion with GADMobileAds.sharedInstance.versionNumber.
+#### [Version 9.6.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/chartboost/ChartboostAdapter-9.6.0.0.zip)
+- Verified compatibility with Chartboost SDK version 9.6.0.
+- Replaced the deprecated method `GADMobileAds.sharedInstance.sdkVersion` with `GADMobileAds.sharedInstance.versionNumber`.
 - Added `GADMediationInterstitialAd` implementation.
+
+Built and tested with:
+- Google Mobile Ads SDK version 10.14.0.
+- Chartboost SDK version 9.6.0.
 
 #### [Version 9.5.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/chartboost/ChartboostAdapter-9.5.1.0.zip)
 - Verified compatibility with Chartboost SDK version 9.5.1.